requirements.txt
Flask==0.12.2
python-binance==0.6.9
coinbase==2.1.0